backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
whitenoise